Use ESLint and fix collision bug
The collision bug was that the boxes for the items being held by the player were only counted as entities part of the room the player was in, so at the beginning when logs are generated, these boxes were only taken into account for collisions in the player's starting room, and in other rooms there was a chance of entities spawing on top of them.

Also added pre-commit hooks to do a minified build, check for TypeScript errors that Parcel may not have picked up on, and run ESLint.

Add .eslintrc.json and README